no more compass
Compass is bringing very little value to the party, especially given that it pulls in a huge amount of dependencies for developers that are otherwise not needed (Ruby, rubygems, bundler, compass, compass-normalize).

We can get the same effects with simpler syntax by just writing clean CSS3 and installing ember-cli-autoprefixer to deal with browser compatibility.